Fix space in project button and add custom component for mobile
